+++ siv3d2025/CreatureBasicParam.h
@@ -16,7 +16,7 @@ namespace GameCore
 						   float attackDelay_secs			 ,
 						   float attackCooldown_secs		 ,
 						   int   attackPower				 ,
-						   int   attackTargetNumber			 ,
+						   float attackTargetNumber			 ,
 						   int   health						 ,
 						   int   defensePower				 ,
 						   float moveSpeed					 ,
@@ -28,14 +28,14 @@ namespace GameCore
 		[[nodiscard]] float			   GetAttackDelay_secs	 () const noexcept { return m_attackDelay_secs;	  }
 		[[nodiscard]] float			   GetAttackCooldown_secs() const noexcept { return m_attackCooldown_secs;}
 		[[nodiscard]] int			   GetAttackPower		 () const noexcept { return m_attackPower;		  }
-		[[nodiscard]] int			   GetAttackTargetNumber () const noexcept { return m_attackTargetNumber; }
+		[[nodiscard]] float			   GetAttackTargetNumber () const noexcept { return m_attackTargetNumber; }
 		[[nodiscard]] int			   GetHealth			 () const noexcept { return m_health;			  }
 		[[nodiscard]] int			   GetDefensePower		 () const noexcept { return m_defensePower;		  }
 		[[nodiscard]] float			   GetMoveSpeed			 () const noexcept { return m_moveSpeed;		  }
 		[[nodiscard]] int			   GetSpawnCost			 () const noexcept { return m_spawnCost;		  }
 		[[nodiscard]] s3d::Vec2		   GetColliderSize		 () const noexcept { return m_colliderSize;		  }
 
-		CreatureBasicParam WithOnDamage(int damage) const; 
+		CreatureBasicParam WithOnDamage(int damage) const;
 		CreatureBasicParam operator+(const CreatureBasicParam& other) const
 		{
 			return CreatureBasicParam{
@@ -59,7 +59,7 @@ namespace GameCore
 		float m_attackDelay_secs	= 10.0f;
 		float m_attackCooldown_secs = 10.0f;
 		int   m_attackPower			= 10;
-		int   m_attackTargetNumber  = 10;
+		float m_attackTargetNumber  = 10;
 		int   m_health				= 10;
 		int   m_defensePower		= 10;
 		float m_moveSpeed			= 10.0f;
